[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-int a, b;
+long long int a, b;
 
 int main(void){
     fastio;
